+++ backend/src/main/java/backend/domain/dto/GameGridRequest.java
@@ -1,4 +1,4 @@
-package backend.dto;
+package backend.domain.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
