+++ backend/src/main/java/backend/service/GameService.java
@@ -3,10 +3,10 @@ package backend.service;
 import backend.domain.Constants;
 import backend.domain.FixedLetter;
 import backend.domain.Language;
+import backend.domain.dto.FixedLetterResponse;
+import backend.domain.dto.GameGridRequest;
+import backend.domain.dto.ValidationResultResponse;
 import backend.domain.entities.Word;
-import backend.dto.FixedLetterResponse;
-import backend.dto.GameGridRequest;
-import backend.dto.ValidationResultResponse;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
