+++ backend/src/test/java/backend/unit/UtilityServiceUnitTests.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import backend.dto.GameGridRequest;
+import backend.domain.dto.GameGridRequest;
 import backend.service.UtilityService;
 import java.util.Arrays;
 import java.util.List;
